Verify whether we can rely on the random permutation properties
We are assuming that the generators with 'nbits' bits per seed,
'4 * nbits' bits per state and 'nbits' bits per returned value
have only '1 / 2^nbits' probability of at least one out of the
total number of '2^nbits' seeds to result in a cycle length shorter
than '2^(2*nbits)'. These probability estimations are assuming
that we have something very closely resembling a random permutation.
Some of the important random permutation properties are explained
at http://www.inference.phy.cam.ac.uk/mackay/itila/ExtraChapters.html

This extra test just nudges the initial PRNG state to derive
different variants of the same generator. Then we check the
number of generators, which result in a cycle length shorter
than '2^(2*nbits)', which is reachable via at least one seed.
The number of tested generator variants is selected in such a
way that we are expecting 32 unlucky generator variants on
average. If the actual number of unlucky generators diverges
too much, then we need to be careful about it.